+++ docs/project/changelog.md
@@ -28,6 +28,10 @@ myst:
 - {{ Fix }} `pyodide config` won't print extra messages anymore.
   {pr}`3483`
 
+- {{ Fix }} Pass the same environment variables for out of tree builds as for in
+  tree builds.
+  {pr}`3495`
+
 - {{ Breaking }} `setStdin` now accepts an extra `autoEOF` parameter. If `true`,
   it will insert an EOF automatically after each string or buffer. Defaults to `true`.
   {pr}`3488`